% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{pepvet_theme_manuscript}
\alias{pepvet_theme_manuscript}
\title{pepVet manuscript theme}
\usage{
pepvet_theme_manuscript()
}
\value{
A ggplot2 theme object.
}
\description{
A compact theme for journal figures. Smaller base text (9pt), thinner grid
lines, tighter margins. Use by adding to any pepVet plot:
\code{plot + pepvet_theme_manuscript()}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  bsa_path <- system.file("extdata", "P02769.fasta", package = "pepVet")
  res <- evaluate_digest(bsa_path, enzyme = "trypsin")
  plot_length_distribution(res) + pepvet_theme_manuscript()
}
}
\seealso{
Other plot-utils:
\code{\link[=pepvet_plot_config]{pepvet_plot_config()}},
\code{\link[=pepvet_plot_config_reset]{pepvet_plot_config_reset()}},
\code{\link[=pepvet_save_figure]{pepvet_save_figure()}},
\code{\link[=pepvet_theme_presentation]{pepvet_theme_presentation()}}
}
\concept{plot-utils}
