% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{pepvet_theme_presentation}
\alias{pepvet_theme_presentation}
\title{pepVet presentation theme}
\usage{
pepvet_theme_presentation()
}
\value{
A ggplot2 theme object.
}
\description{
A bold theme for talks and posters. Larger base text (14pt), thicker grid
lines, wider margins. Use by adding to any pepVet plot:
\code{plot + pepvet_theme_presentation()}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  bsa_path <- system.file("extdata", "P02769.fasta", package = "pepVet")
  res <- evaluate_digest(bsa_path, enzyme = "trypsin")
  plot_length_distribution(res) + pepvet_theme_presentation()
}
}
\seealso{
Other plot-utils:
\code{\link[=pepvet_plot_config]{pepvet_plot_config()}},
\code{\link[=pepvet_plot_config_reset]{pepvet_plot_config_reset()}},
\code{\link[=pepvet_save_figure]{pepvet_save_figure()}},
\code{\link[=pepvet_theme_manuscript]{pepvet_theme_manuscript()}}
}
\concept{plot-utils}
