% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepVet-package.R
\docType{package}
\name{pepVet-package}
\alias{pepVet}
\alias{pepVet-package}
\title{pepVet: Evaluate Proteolytic Digests for Proteomics Workflows}
\description{
Simulates proteolytic digestion, scores peptides for LC-MS/MS suitability,
compares candidate enzymes, and reports digest quality at the protein level.
Includes validated cleavage rules for 40 enzymes, workflow presets, peptide
mass and pI utilities, batch evaluation, and styled console reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LangeLab/pepVet}
  \item \url{https://langelab.github.io/pepVet/}
  \item Report bugs at \url{https://github.com/LangeLab/pepVet/issues}
}

}
\author{
\strong{Maintainer}: Enes K. Ergin \email{eneskemalergin@gmail.com} (\href{https://orcid.org/0000-0001-9810-7399}{ORCID})

Authors:
\itemize{
  \item Enes K. Ergin \email{eneskemalergin@gmail.com} (\href{https://orcid.org/0000-0001-9810-7399}{ORCID})
}

}
\keyword{internal}
