% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepVet-package.R
\docType{package}
\name{pepVet-package}
\alias{pepVet}
\alias{pepVet-package}
\title{pepVet: Evaluate Proteolytic Digests for Proteomics Workflows}
\description{
\code{pepVet} provides tools for proteolytic digest simulation, peptide-set scoring, enzyme comparison, workflow preset selection, and console reporting for bottom-up proteomics.
}
\details{
Version 0.1.0 completes the Phase 2 scoring and annotation pass with workflow presets, peptide mass and pI utilities, preset-tracking metadata, and sequence-local cleavage-efficiency annotations for trypsin-family digests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LangeLab/pepVet}
  \item \url{https://langelab.github.io/pepVet/}
  \item Report bugs at \url{https://github.com/LangeLab/pepVet/issues}
}

}
\author{
\strong{Maintainer}: Enes K. Ergin \email{eneskemalergin@gmail.com}

Authors:
\itemize{
  \item Enes K. Ergin \email{eneskemalergin@gmail.com}
}

}
\keyword{internal}
