% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepVet-package.R
\docType{package}
\name{pepVet-package}
\alias{pepVet}
\alias{pepVet-package}
\title{pepVet: Evaluate Proteolytic Digests for Proteomics Workflows}
\description{
\code{pepVet} provides tools for simulating proteolytic digestion and evaluating
peptide sets for proteomics workflows.
}
\details{
The current package version provides validated digest generation,
amino-acid reference data, and a core protein-level scoring engine built for
iterative proteomics workflow development on R 4.5 and Bioconductor 3.22.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LangeLab/pepVet}
  \item \url{https://langelab.github.io/pepVet/}
  \item Report bugs at \url{https://github.com/LangeLab/pepVet/issues}
}

}
\author{
\strong{Maintainer}: Enes K. Ergin \email{eneskemalergin@gmail.com}

}
\keyword{internal}
