% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepVet-package.R
\docType{package}
\name{pepVet-package}
\alias{pepVet}
\alias{pepVet-package}
\title{pepVet: Evaluate Proteolytic Digests for Proteomics Workflows}
\description{
\code{pepVet} provides tools for proteolytic digest simulation, peptide-set scoring, enzyme comparison, workflow preset selection, and console reporting for bottom-up proteomics.
}
\details{
Version 0.0.4 adds configurable peptide-length and GRAVY scoring windows, workflow presets for common acquisition settings, enzyme-aware peptide-count normalisation, and user-facing articles that explain how to apply the package in real experimental settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LangeLab/pepVet}
  \item \url{https://langelab.github.io/pepVet/}
  \item Report bugs at \url{https://github.com/LangeLab/pepVet/issues}
}

}
\author{
\strong{Maintainer}: Enes K. Ergin \email{eneskemalergin@gmail.com}

}
\keyword{internal}
